+++ spinn_utilities/citation/__init__.py
@@ -1,4 +1,4 @@
-from .citation_and_doi_updater import CitationUpdaterAndDoiGenerator
+from .citation_updater_and_doi_generator import CitationUpdaterAndDoiGenerator
 from .tool_citation_generation import CitationAggregator
 from .generate_aggregate import generate_aggregate
