+++ spinn_utilities/citation/citation_updater_and_doi_generator.py
